examples/frontend/requirements.txt
torch==1.8.1
transformers==3.3.1
sentence-transformers==0.3.8
pandas==1.1.2
faiss-cpu==1.6.1
faiss-gpu
numpy==1.19.2
folium==0.2.1
streamlit==0.62.0
plotly
sklearn
Pillow